+++ sound-mind-studio/tests/test_layers_panel.h
@@ -11,6 +11,8 @@ private slots:
     void visibilityButtonEmitsVisibilityToggled();
     void backgroundVisibilityButtonIsDisabled();
     void opacitySliderEmitsOpacityChanged();
+    void translationSpinBoxEmitsTranslationChanged();
+    void rescaleSpinBoxEmitsRescaleChanged();
     void doubleClickingNameEmitsRenameRequested();
     void deleteButtonEmitsDeleteRequestedForNormalLayers();
     void lockedLayersHaveNoDeleteButton();