+++ sound-mind-studio/tests/test_main_window.h
@@ -116,4 +116,8 @@ private slots:
     void handleDroppedFilesIgnoresUnrecognizedExtensions();
     void handleDroppedFilesRoutesMultipleFilesInOrder();
     void handleDroppedFilesSmprojRefusesWhileLoopModeIsRunning();
+
+    // Layer Time Alignment (v0.Y.21.1)
+    void setLayerTranslationChangesTheLayersTranslation();
+    void setLayerRescaleChangesTheLayersRescale();
 };